Button: enhancements with ButtonLink and IconButton (#1659)
* Create icon only button

* Add ButtonLink and IconButton components

* Lint and test

* Remove react router for now. Maybe add an example later

* Fix the typescript error on as prop

* Fix lint issue

* Fix
1. Bug with disabled
2. Tests in testing lib

* lint issues
